# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/categories": {
      "post": {
        "description": "Adds a category to a group of the active budget, at the end of that group.",
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The category that was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "The body was refused, or the change names something the caller cannot reach. A refusal from the domain carries the reason it was refused for.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRefusedResponse"
                }
              }
            }
          },
          "401": {
            "description": "The token was missing, malformed, expired or not minted for this app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "409": {
            "description": "The idempotency key was claimed by a different request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponse"
                }
              }
            }
          }
        },
        "summary": "Create a category",
        "tags": [
          "Categories"
        ]
      }
    },
    "/categories/reorder": {
      "post": {
        "description": "Takes the categories of the group in the order the user left them. The list may name fewer than the group holds, because the month it came from lists only what is visible in that month, and the rest keep their order behind the named ones. A duplicate or an id the group does not hold is refused.",
        "operationId": "CategoriesController_reorder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderCategoriesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The categories in their new order.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "The body was refused, or the change names something the caller cannot reach. A refusal from the domain carries the reason it was refused for.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRefusedResponse"
                }
              }
            }
          },
          "401": {
            "description": "The token was missing, malformed, expired or not minted for this app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "409": {
            "description": "The idempotency key was claimed by a different request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponse"
                }
              }
            }
          }
        },
        "summary": "Rewrite the order of one group",
        "tags": [
          "Categories"
        ]
      }
    },
    "/categories/{id}": {
      "patch": {
        "description": "Every field is optional and only what is sent is written. Moving a category into a hidden group is refused: the money it holds would leave the screen with it.",
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The category as it stands now.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "The body was refused, or the change names something the caller cannot reach. A refusal from the domain carries the reason it was refused for.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRefusedResponse"
                }
              }
            }
          },
          "401": {
            "description": "The token was missing, malformed, expired or not minted for this app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "409": {
            "description": "The idempotency key was claimed by a different request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponse"
                }
              }
            }
          }
        },
        "summary": "Rename a category, move it or change its look",
        "tags": [
          "Categories"
        ]
      }
    },
    "/categories/{id}/hide": {
      "post": {
        "description": "Marks the category hidden from this moment on, which takes it out of every month after that one and out of none before it. The row and its transactions are untouched, so its past activity keeps counting. A category holding money over any month is refused.",
        "operationId": "CategoriesController_hide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdempotentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The category, now hidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "The body was refused, or the change names something the caller cannot reach. A refusal from the domain carries the reason it was refused for.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRefusedResponse"
                }
              }
            }
          },
          "401": {
            "description": "The token was missing, malformed, expired or not minted for this app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "409": {
            "description": "The idempotency key was claimed by a different request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponse"
                }
              }
            }
          }
        },
        "summary": "Hide a category",
        "tags": [
          "Categories"
        ]
      }
    },
    "/categories/{id}/unhide": {
      "post": {
        "description": "Takes the hidden marker off, and the category is in every month again.",
        "operationId": "CategoriesController_unhide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdempotentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The category, visible again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "The body was refused, or the change names something the caller cannot reach. A refusal from the domain carries the reason it was refused for.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRefusedResponse"
                }
              }
            }
          },
          "401": {
            "description": "The token was missing, malformed, expired or not minted for this app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "409": {
            "description": "The idempotency key was claimed by a different request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponse"
                }
              }
            }
          }
        },
        "summary": "Bring a hidden category back",
        "tags": [
          "Categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Rondo Money API",
...
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "clerkSessionToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "A Clerk session token. In the web app: `await getToken()`."
      }
    },
    "schemas": {
...
      "UnauthorizedResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "Unauthorized"
          },
          "message": {
            "type": "string",
            "example": "Invalid session token"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "ConflictResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "error": {
            "type": "string",
            "example": "Conflict"
          },
          "message": {
            "type": "string",
            "example": "Refusing the mutation carrying key \"form-opened-once\": the idempotency key was claimed by a different request, and answering with the first one's result would report a write it never made"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "CategoryIcon": {
        "type": "string",
        "enum": [
          "home",
          "bolt",
          "droplet",
          "flame",
          "wifi",
          "phone",
          "tool",
          "sofa",
          "wash",
          "shopping-cart",
          "restaurant",
          "coffee",
          "pizza",
          "beer",
          "car",
          "gas-station",
          "bus",
          "parking",
          "bike",
          "plane",
          "beach",
          "heartbeat",
          "pill",
          "stethoscope",
          "dental",
          "barbell",
          "shirt",
          "scissors",
          "perfume",
          "mood-smile",
          "baby-carriage",
          "school",
          "book",
          "paw",
          "gift",
          "tv",
          "movie",
          "gamepad",
          "music",
          "ticket",
          "repeat",
          "cloud",
          "laptop",
          "pig-money",
          "credit-card",
          "receipt-tax",
          "shield-check",
          "building-bank",
          "heart-handshake",
          "tag"
        ],
        "description": "Which icon this category is drawn with, as a domain name rather than the name of a component. A category nobody has given one carries null, and so does a stored name this app no longer draws."
      },
      "CategoryColor": {
        "type": "string",
        "enum": [
          "scarlet",
          "vermilion",
          "orange",
          "amber",
          "gold",
          "yellow",
          "lime",
          "green",
          "emerald",
          "mint",
          "teal",
          "cyan",
          "sky",
          "blue",
          "indigo",
          "violet",
          "purple",
          "orchid",
          "magenta",
          "pink",
          "rose",
          "brown",
          "graphite",
          "gray"
        ],
        "description": "Which colour this category is drawn in, on the same terms as its icon."
      },
...
      "CategoryRefusal": {
        "type": "string",
        "enum": [
          "ALREADY_HIDDEN",
          "AVAILABLE_NOT_ZERO",
          "GROUP_HIDDEN",
          "NO_ACTIVE_BUDGET",
          "UNKNOWN_CATEGORY",
          "UNKNOWN_GROUP"
        ],
        "description": "Why the change was refused, for a screen that answers each refusal differently rather than by reading the message. It is absent when the body itself was refused, because the pipe answers before the domain has a reason to give."
      },
      "CategoryRefusedResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "name must be longer than or equal to 1 characters"
            ]
          },
          "reason": {
            "description": "Why the change was refused, for a screen that answers each refusal differently rather than by reading the message. It is absent when the body itself was refused, because the pipe answers before the domain has a reason to give.",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryRefusal"
              }
            ]
          },
          "available": {
            "type": "string",
            "pattern": "^(0|-?[1-9]\\d*)$",
            "maxLength": 20,
            "example": "-4500",
            "description": "What the category still holds over every month, present only when that is what blocked the hide. The screen names this amount rather than the one the open month shows."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "IdempotentDto": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "description": "Minted once when the form opens, never per request. A key per request makes a double click two writes again.",
            "minLength": 1,
            "maxLength": 64
          }
        },
        "required": [
          "idempotencyKey"
        ],
        "additionalProperties": false
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "description": "Minted once when the form opens, never per request. A key per request makes a double click two writes again.",
            "minLength": 1,
            "maxLength": 64
          },
          "groupId": {
            "type": "string",
            "format": "uuid",
            "description": "The group the category is created in."
          },
          "name": {
            "type": "string",
            "description": "What the user calls this category.",
            "maxLength": 60
          },
          "icon": {
            "example": "home",
            "nullable": false,
            "description": "Which icon it is drawn with.",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryIcon"
              }
            ]
          },
          "color": {
            "example": "blue",
            "nullable": false,
            "description": "Which colour it is drawn in.",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryColor"
              }
            ]
          }
        },
        "required": [
          "idempotencyKey",
          "groupId",
          "name"
        ],
        "additionalProperties": false
      },
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "groupId": {
            "type": "string",
            "format": "uuid",
            "description": "The group this category sits in."
          },
          "name": {
            "type": "string",
            "description": "What the user calls this category."
          },
          "sortOrder": {
            "type": "number",
            "description": "Where the category sits in its group, counted from zero."
          },
          "icon": {
            "example": "home",
            "nullable": true,
            "description": "Which icon this category is drawn with, or null when nobody picked one.",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryIcon"
              }
            ]
          },
          "color": {
            "example": "blue",
            "nullable": true,
            "description": "Which colour this category is drawn in, on the same terms as its icon.",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryColor"
              }
            ]
          },
          "hidden": {
            "type": "boolean",
            "description": "Whether the category carries a hidden marker at all."
          }
        },
        "required": [
          "id",
          "groupId",
          "name",
          "sortOrder",
          "icon",
          "color",
          "hidden"
        ]
      },
      "ReorderCategoriesDto": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "description": "Minted once when the form opens, never per request. A key per request makes a double click two writes again.",
            "minLength": 1,
            "maxLength": 64
          },
          "groupId": {
            "type": "string",
            "format": "uuid",
            "description": "The group whose order is being rewritten."
          },
          "categoryIds": {
            "description": "The categories of that group, in the order the user left them. Fewer than the group holds is allowed; the rest keep their order behind the named ones.",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "idempotencyKey",
          "groupId",
          "categoryIds"
        ],
        "additionalProperties": false
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "description": "Minted once when the form opens, never per request. A key per request makes a double click two writes again.",
            "minLength": 1,
            "maxLength": 64
          },
          "groupId": {
            "type": "string",
            "format": "uuid",
            "description": "The group to move it into."
          },
          "name": {
            "type": "string",
            "maxLength": 60,
            "description": "What to call it now."
          },
          "icon": {
            "example": "home",
            "nullable": false,
            "description": "Which icon it is drawn with.",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryIcon"
              }
            ]
          },
          "color": {
            "example": "blue",
            "nullable": false,
            "description": "Which colour it is drawn in.",
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryColor"
              }
            ]
          }
        },
        "required": [
          "idempotencyKey"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "security": [
    {
      "clerkSessionToken": []
    }
  ]
}